requirements.txt
# PyTorch with CUDA support
# Platform-specific installation handled in README
torch>=2.0.0
torchaudio>=2.0.0

# Core speech recognition
openai-whisper==20231117

# Audio processing
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.4

# Terminal UI and utilities
colorama==0.4.6
rich==13.7.0

# System interaction
keyboard==0.13.5
psutil==5.9.6

# GPU monitoring
pynvml==11.5.0

# Audio format support
librosa==0.10.1
soundfile==0.12.1

# Threading and async
asyncio-throttle==1.0.2